split outbrain into three modules for easier testing and no default exports
